csv_1.py
import streamlit as st 
from gemini_api import geminiapi  # Placeholder for Gemini API import
from dotenv import load_dotenv
import os
import pandas as pd
from pandasai import SmartDataframe

load_dotenv()

gemini_api_key = os.getenv("GEMINI_API_KEY")  # Placeholder for Gemini API key

def chat_with_csv(df, prompt):
    print("Point 1")
    gemini = geminiapi(api_key=gemini_api_key)  # Placeholder for Gemini API initialization
    print("Point 2")
    # Use GeminiAPI methods to process natural language queries
    print("Point 3")
    # Process prompt using GeminiAPI
    result = "Placeholder result from GeminiAPI"  # Placeholder for Gemini API result
    print("Point 4")
    print(result)   
    return result

st.set_page_config(layout='wide')
st.title("ChatCSV powered by Gemini API")  # Update title to reflect Gemini API

input_csv = st.file_uploader("Upload your CSV file", type=['csv'])

if input_csv is not None:
    col1, col2 = st.columns([1, 1])

    with col1:
        st.info("CSV Uploaded Successfully")
        data = pd.read_csv(input_csv)
        st.dataframe(data, use_container_width=True)

    with col2:
        st.info("Chat Below")
        input_text = st.text_area("Enter your query")

        if input_text is not None:
            if st.button("Chat with CSV"):
                st.info("Your Query: " + input_text)
                result = chat_with_csv(data, input_text)
                st.success(result)
