requirements.txt
Flask
langchain
langchain-google-genai
pandas
python-dotenv
PyPDF2
streamlit
pandasai